Live-2-Vod Cleanup
This commit cleans the live2vod feature. Note that there are no changes
to the text. What I did in this commit was a general cleanup to remove
any bikeshed warnings and errors.

I also aligned formatting of the file with a hard wrap at 80 characters.
The 80 characters is just a start, and we can adjust this, but enforcing
hard wraps is very helpful to get better git diffs on changes since git
diffs in full lines.

I also cleaned up some of the formatting to make it more consistent.
This is something that we can review later of course.