Bikeshed builds (#425)
This squashed commit bring along a new build pipeline to automate the bikeshed builds from a docker container and quite a bit of cleanup around the overall structure of the repository. There is still some work to do, especially around the documentation how to use bikeshed as well as the container. We also need some easy to use build scripts for windows and a few other details, but this should be a reasonable start.

What happened in this commit was that I
 - cleaned the bikeshed source files and updated the repo structure to use sub-folder pear feature spec
 - adopted the documents to be compatible with the latest bikeshed
 - created the build container and included some dashif specific boilerplate 
 - moved the references to `biblio.json` since this will be automatically picked up when we run bikeshed
 - use a remote link for the logo and moved the snippet to the boilerplate
 - added vscode and editor config for easier integration with VSCode and other editors
 - cleaned Live-2-Vod: Note that there are no changes to the text. What I did in this commit was a general cleanup to
    remove any bikeshed warnings and errors.
 - more cleanup for general formatting
 - added a GH Actions based pipeline for the container, the master branch as well as PR builds. This publishes to GH Pages

I also aligned formatting of the file with a hard wrap at 80 characters. The 80 characters is just a start, and we can adjust this, but enforcing hard wraps is very helpful to get better git diffs on changes since git diffs in full lines.