Add support for single spec repos
There are cases where we want to use the docker image that is created here in
other repositories to compile a spec. They do not follow the structure implied
here and rather contain a single bikeshed file.

This commit adds support for this in the Makefile and the spec source and its
name can be set using SRC and NAME variables.

The output folder will be a flat dist/ folder.

Note that for now we still assume that images are located in the Images folder
and this is not configurable at the moment.