Added plantuml Diagra support
This commit patches the build tools and adds support to compile plantuml
diagrams from wsd files in the Diagrams folder. This works for single specs
that are located in the root folder as well as for nested structures in the
specs folder.
Note that nesting with sub-folders inside the Diagrams folder is not supported.