Fixed indentation and missing quite
Apparently bikesheds markdown interpreter is _very_ picky with indentations
for lists. In the last edit there was a _tab_ and two _space_ characters in
two lines. This probably looked properly aligned in the editor, but it was not
for bikeshed.

There was also a quite missing, which messed up linking and was also reported
as an issue.

I fixed it now here and will also put something into the authoring document to
indicated that this can be a problem.