Fix bikeshed build
We are bringing our own spec defaults and install them into the
bikeshed folder when building our container. Now bikeshed requires
an entry in the KDF dataset in order to find the group that we defined.

This commit adds the KDF entry to the list and also defines the
"CR", i.e. "Community Review" status. This is a reasonable start for a
more advanced status management for the documents we maintain here.